+++ src/components/NotificationComponent.vue
@@ -30,28 +30,3 @@ export default {
   },
 };
 </script>
-
-<style scoped>
-.notification {
-  position: absolute;
-  top: 40%; /* Display at the bottom */
-  left: 50%;
-  transform: translateX(-50%);
-  padding: 10px 20px;
-  border-radius: 5px;
-  color: white;
-  text-align: center;
-  opacity: 1;
-  transition: opacity 0.5s;
-  margin-bottom: 20px; /* Add space between notification and input */
-  z-index: 1; /* Ensure the notification appears above other content */
-}
-
-.notification.success {
-  background-color: #28a745; /* Green background for success */
-}
-
-.notification.error {
-  background-color: #dc3545; /* Red background for error */
-}
-</style>