+++ src/main.js
@@ -6,6 +6,8 @@ import { library } from '@fortawesome/fontawesome-svg-core';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 import { fas } from '@fortawesome/free-solid-svg-icons';
 
+import "./assets/scss/styles.scss";
+
 library.add(fas);
 
 // Use app.component instead of Vue.component